Update README.md
NOTE:  Earlier commits were made to a previous version of the notebook. " Assessment_Weather_App_Final_ipynb.ipynb" which is part of the REPO.
Please refer to the file `` for the ** WEATHER ADVISOR APP"* of the project. This final version includes the complete interactive weather dashboard, updated data source, and all functionality described in the submission requirements.